Reverted some changes in ^ and add new tests
I reverted some changes in ^ with FloatinPoint or Interval powers.
The criteria to implement the simplest algorithm is if x, the power,
is a thin interval, in the sense that its width is at most eps(x).
So I reverted to include the comparison of the width with eps(x) when
x is a FloatingPoint, and set the maximum width to eps(mid(x)) when
x is an interval.

I added some tests using x=1/3, and now all return the same interval.